+++ src/fr/univmrs/tagc/GINsim/export/regulatoryGraph/GsPetriNetExportAPNN.java
@@ -48,7 +48,7 @@ public class GsPetriNetExportAPNN extends GsPetriNetExport
         int len = v_no.size();
         OmddNode[] t_tree = ((GsRegulatoryGraph)graph).getAllTrees(true);
         List[] t_transition = new List[len];
-        short[][] t_markup = GsPetriNetExport.prepareExport(config, t_transition, t_tree);
+        byte[][] t_markup = GsPetriNetExport.prepareExport(config, t_transition, t_tree);
 
         try 
         {
